Extract particle collection to separate file
